requirements.txt
# Core Discord Bot
discord.py[voice] @ git+https://github.com/rapptz/discord.py
python-dotenv>=1.0.0

# Audio Processing
yt-dlp>=2025.1.1
PyNaCl>=1.5.0
aiohttp>=3.9.0
pydub>=0.25.1

# Database & ORM
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0  # PostgreSQL async driver
aiosqlite>=0.20.0  # SQLite async driver

# Caching
redis>=5.0.0

# Spotify Integration
spotipy>=2.24.0

# Audio Filters & Processing
numpy>=1.24.0
scipy>=1.10.0

# Utilities
asyncio-throttle>=1.0.0
dataclasses-json>=0.6.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Monitoring & Logging
structlog>=24.0.0
